Default return value for Piper#gets.
The default return value is used when the gets method times out or when no
data is returned from the read call on the unix socket. This is great for
detecting timeouts and error conditions.

Also, now using a Socket#pair instead of a pipe for bi-drectional
communication. The author was confused about the operation of pipes.