Reverting back to blocking IO.
For tight execution loops, the non-blocking IO was causing a loss of data from
the pipe -- a read would happen before all the data was written to the socket
from the other end. Since we are reading fixed sizes of data, this was causing
an error to be raised on read.