Wrapped the run loop in a catch block for quick exits.
A catch block is being used to quickly exit the run loop when needed. This
allows your code to respond to stop requests much more easily - simply throw
:halt_run_loop.

Beware because this will bypass the all the code following the throw call and
exit the run loop immediately. The after_stopping block will be run.

Example:

  def run
    ...
    throw :halt_run_loop unless running?
    ..
  end