Fix the alive? check in the Child class
We check if the child process is alive by sending a kill -0 signal to the
child. If this method returns without error, then the child is alive and well.
Except in the case where the child has terminated and the PID has not yet been
reaped by the parent. The simple fix is to wait(WNOHANG) on the child PID
beforing checking the state of the child.