Add strict interval semantics to the threaded class
A strict interval is used to ensure that subsequent executions of the run loop
occur as close to "interval" seconds as is possible. Currently the run loop
executions drift by the amount of time it takes to run the code. The strict
interval semantics take this time into account when sleeping between
executions.

addresses #9